+++ src/lib/containers/client/Step/Step.tsx
@@ -1,9 +1,10 @@
 "use client";
 
-import { PropsWithChildren, createContext, useState } from "react";
+import { PropsWithChildren, createContext } from "react";
 
 import { StepContextValue } from "./Step.type";
 import { StepContent } from "./components";
+import { useStepControl } from "./hooks";
 
 export const StepContext = createContext<StepContextValue>({
   step: 0,
@@ -13,30 +14,9 @@ export const StepContext = createContext<StepContextValue>({
 });
 
 const Step = ({ children }: PropsWithChildren) => {
-  const [step, setStep] = useState(0);
-  const [maxStep, setMaxStep] = useState(0);
+  const stepController = useStepControl();
 
-  const prevStep = () =>
-    setStep(prev => {
-      if (prev === 0) return prev;
-
-      return prev - 1;
-    });
-
-  const nextStep = () =>
-    setStep(prev => {
-      if (prev === maxStep) return prev;
-
-      return prev + 1;
-    });
-
-  const handleMaxStep = (maxCount: number) => setMaxStep(maxCount);
-
-  return (
-    <StepContext.Provider value={{ step: step, nextStep, prevStep, setMaxStep: handleMaxStep }}>
-      {children}
-    </StepContext.Provider>
-  );
+  return <StepContext.Provider value={stepController}>{children}</StepContext.Provider>;
 };
 
 Step.Content = StepContent;
